+++ src/cuid.rs
@@ -34,20 +34,12 @@ pub fn parse_cuid1(args: &Args) -> Option<IDInfo> {
         id_type: "CUID".to_string(),
         version: Some("1".to_string()),
         standard: args.id.to_string(),
-        integer: None,
-        short_uuid: None,
-        base64: None,
-        uuid_wrap: None,
-        size: 0,
-        entropy: 0,
         datetime: Some(datetime),
         timestamp: Some(timestamp.to_string()),
         sequence: Some(sequence as u128),
         node1: Some(format!("{} (Fingerprint)", fingerprint)),
         node2: Some(format!("{} (Random data)", random_data)),
-        hex: None,
-        bits: None,
-        color_map: None,
+        ..Default::default()
     })
 }
 
@@ -62,19 +54,6 @@ pub fn parse_cuid2(args: &Args) -> Option<IDInfo> {
         id_type: "CUID".to_string(),
         version: Some("2".to_string()),
         standard: args.id.to_string(),
-        integer: None,
-        short_uuid: None,
-        base64: None,
-        uuid_wrap: None,
-        size: 0,
-        entropy: 0,
-        datetime: None,
-        timestamp: None,
-        sequence: None,
-        node1: None,
-        node2: None,
-        hex: None,
-        bits: None,
-        color_map: None,
+        ..Default::default()
     })
 }