+++ src/objectid.rs
@@ -21,24 +21,19 @@ pub fn parse_objectid(args: &Args) -> Option<IDInfo> {
 
     Some(IDInfo {
         id_type: "MongoDB ObjectId".to_string(),
-        version: None,
         standard: args.id.to_string(),
         integer: Some(oid_int),
-        short_uuid: None,
-        base64: None,
-        uuid_wrap: None,
         size: 96,
         entropy: 40,
         datetime: Some(datetime),
         timestamp: Some(timestamp),
         sequence: Some(sequence),
-        node1: None,
-        node2: None,
         hex: Some(hex::encode(oid_bytes.clone())),
         bits: Some(oid_bytes.iter().fold(String::new(), |mut output, c| {
             let _ = write!(output, "{c:08b}");
             output
         })),
         color_map: Some("333333333333333333333333333333332222222222222222222222222222222222222222666666666666666666666666".to_string()),
+        ..Default::default()
     })
 }