+++ src/timeflake.rs
@@ -16,25 +16,20 @@ pub fn parse_timeflake_core(hex_id: &str) -> Option<IDInfo> {
 
     Some(IDInfo {
         id_type: "Timeflake".to_string(),
-        version: None,
         standard: format!("{:0>22}", base62::encode(timeflake.as_u128())),
         integer: Some(timeflake.as_u128()),
-        short_uuid: None,
-        base64: None,
         uuid_wrap: Some(timeflake.as_uuid().to_string()),
         size: 128,
         entropy: 80,
         datetime: Some(datetime),
         timestamp: Some(timestamp.to_string()),
-        sequence: None,
-        node1: None,
-        node2: None,
         hex: Some(hex_id.to_string()),
         bits: Some(timeflake.as_u128().to_be_bytes().iter().fold(String::new(), |mut output, c| {
             let _ = write!(output, "{c:08b}");
             output
         })),
         color_map: Some("33333333333333333333333333333333333333333333333322222222222222222222222222222222222222222222222222222222222222222222222222222222".to_string()),
+        ..Default::default()
     })
 }