+++ CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.2.3] - 2024-12-21
+
+### Added
+
+- Hashids support.
+- YouTube Video ID support.
+- Stripe ID support.
+
+### Changed
+
+- Fix security issue from crate time.
+
 ## [0.2.2] - 2024-12-10
 
 ### Changed